Codegen.Identifiers

Description

Identifiers in generated C code and helper functions for constructing them.

Documentation

newtype CIdent

Use snake_case for c literals

Identifiers in C.

Constructors

CIdent Identifier

ctype :: CIdent -> Type

Construct a type name from a C identifier.

cexpr :: CIdent -> Exp

Construct an expression from a C identifier.

time_t :: Type

C type that represents model time

priority_t :: Type

C type that represents process priorities

depth_t :: Type

C type that represents process depths

trigger_t :: Type

C type that represents triggers, aka processes that are sensitized on variables

stepf_t :: Type

C type that represents the step function of a process

bool_t :: Type

C type that represents booleans

act_t :: Type

The type of the activation record base class.

pc :: CIdent

Name of the program counter field in an act_t.

caller :: CIdent

Name of the caller field in an act_t.

depth :: CIdent

Name of the depth field in an act_t.

priority :: CIdent

Name of the priority field in an act_t.

sv_t :: Type

The type of the scheduled variable base class.

word_t :: Type

The word size as type uint32_t

value :: CIdent

The conventional name of the payload field in a type-specialized SV.

container_of :: CIdent

Name of the pseudonymous macro from the Linux kernel.

initialize_program :: CIdent

Name of top level program initialization function

initialize_static_input_device :: CIdent

Name of the top level static input switch initialization function

initialize_static_output_device :: CIdent

Name of the top level static output initialization function

top_return :: CIdent

Name of top level return step-function

top_parent :: CIdent

Name of top level parent activation record

activate :: CIdent

Name of routine that forks procedures

enter :: CIdent

Name of routine that initialized an activation record

leave :: CIdent

Name of routine that deallocates an activation record

event_on :: CIdent

Name of routine that checks if a reference has been written to

sensitize :: CIdent

Name of routine that sensitizes a procedure

desensitize :: CIdent

Name of routine that desensitizes a procedure

unsched_event :: CIdent

Name of routine that dequeues an event from the event queue

now :: CIdent

Name of routine that returns the current value of now.

never :: CIdent

Name of routine that returns the current value of now.

root_priority :: CIdent

Priority of the root SSM routine.

root_depth :: CIdent

Depth of the root SSM routine.

throw :: CIdent

Name of macro that throws an error.

exhausted_priority :: CIdent

Exhausted priority

debug_trace :: CIdent

Insert debug trace

debug_microtick :: CIdent

Used for inserting

act_ :: VarId -> CIdent

Obtain the name of the activation record struct for a routine.

step_ :: VarId -> CIdent

Obtain the name of the step function of a routine.

enter_ :: VarId -> CIdent

Obtain the name for the enter function of a routine.

trig_ :: Int -> CIdent

Obtain the name of each trigger for a routine.

tmp_ :: Int -> CIdent

Obtain the name of a temporary variable.

arg_ :: Int -> CIdent

Obtain the name of an argument variable.

leave_cleanup :: CIdent

Label in the step function where cleanup is performed before leaving.

acts :: CIdent

Identifier for specialized (outer) struct act.

actg :: CIdent

Identifier for generic (inner) struct act.

act_member :: CIdent

Identifier for act member in act struct.

ret_val :: CIdent

Name of return argument.

sv_ :: Identifiable a => a -> CIdent

Name of the scheduled variable type for an SSM Type.

initialize_ :: Identifiable a => a -> CIdent

Name of the initialize method for an SSM Type.

assign_ :: Identifiable a => a -> CIdent

Name of the assign method for an SSM Type.

later_ :: Identifiable a => a -> CIdent

Name of the later method for an SSM Type.

sv :: CIdent

Name of the generic ssm_sv_t embedded inside of a scheduled variable type.

int_ :: Int -> CIdent

Name of C integer type.

mem_alloc :: CIdent

Allocate memory.

mem_free :: CIdent

Free memory.