Codegen.LibSSM

Description

Bindings to the ssm-runtime library.

Documentation

newtype CIdent

Identifiers in C.

Constructors

CIdent Identifier

ctype :: CIdent -> Type

Construct a type name from a C identifier.

cexpr :: CIdent -> Exp

Construct an expression from a C identifier.

csizeof :: Type -> Exp

Construct an expression of the size of a C type.

ccall :: Exp -> [Exp] -> Exp

amp :: Exp -> Exp

star :: Exp -> Exp

data SSMError

enum ssm_error, an enumeration of possible runtime errors.

Constructors

INTERNAL_ERROR
Reserved for unforeseen, non-user-facing errors.

EXHAUSTED_ACT_QUEUE
Tried to insert into full ready queue.

EXHAUSTED_EVENT_QUEUE
Tried to insert into full event queue.

EXHAUSTED_MEMORY
Could not allocate more memory.

EXHAUSTED_PRIORITY
Tried to exceed available recursion depth.

NOT_READY
Not yet ready to perform the requested action.

INVALID_TIME
Specified invalid time.

INVALID_MEMORY
Invalid memory layout.

throw :: SSMError -> Exp

SSM_THROW, throw a runtime error.

word_t :: Type

ssm_word_t, the largest supported machine word size.

value_t :: Type

ssm_value_t, runtime representation for sslang values.

packed_val :: CIdent

Name of the packed_val field in a value_t.

heap_ptr :: CIdent

Name of the heap_ptr field in a value_t.

mm_tag :: CIdent

marshal :: Exp -> Exp

ssm_marshal, construct a value_t out of a 31-bit integral value.

unmarshal :: Exp -> Exp

ssm_unmarshal, extract 31-bit integral value out of a value_t.

on_heap :: Exp -> Exp

ssm_on_heap, whether a value_t points to something on the heap.

dup :: Exp -> Exp

ssm_dup, increment the reference count of a value.

drop :: Exp -> Exp

ssm_drop, drop the reference count of a value and free if necessary.

time_t :: Type

ssm_time_t, 64-bit time type.

time_obj_t :: Type

struct ssm_time, the heap-representation of a time_t.

never :: Exp

SSM_NEVER, a time_t that is never reached.

now :: Exp

ssm_now, the current time.

new_time :: Exp -> Exp

ssm_new_time, allocate a time_obj_t on the heap.

read_time :: Exp -> Exp

ssm_time_read, read the time in a time_obj_t.

priority_t :: Type

ssm_priority_t, thread priority.

depth_t :: Type

ssm_depth_t, thread depth.

root_priority :: Exp

SSM_ROOT_PRIORITY, the depth of the root process.

root_depth :: Exp

SSM_ROOT_DEPTH, the depth of the root process.

act_t :: Type

ssm_act_t, the generic activation record type.

act_pc :: CIdent

Name of the program counter field in an act_t.

act_caller :: CIdent

Name of the caller field in an act_t.

act_depth :: CIdent

Name of the depth field in an act_t.

act_priority :: CIdent

Name of the priority field in an act_t.

trigger_t :: Type

ssm_trigger_t, nodes in the linked list of triggers.

enter :: Exp -> Exp -> Exp -> Exp -> Exp -> Exp

ssm_enter, allocate and initialize activation record.

leave :: Exp -> Exp -> Exp

ssm_leave, deallocate an activation record before leaving.

activate :: Exp -> Exp

ssm_activate, schedule an activation record on the ready queue.

top_parent :: Exp

ssm_top_parent, Name of top level parent activation record

sv_t :: Type

ssm_sv_t, polymorphic scheduled variables.

new_sv :: Exp -> Exp

ssm_new_sv, allocate a new sv_t on the heap.

to_sv :: Exp -> Exp

ssm_to_sv, retrieve the sv_t pointer pointed to by a value_t.

deref :: Exp -> Exp

ssm_deref, read the value of an sv_t pointed to by a value_t.

assign :: Exp -> Exp -> Exp -> Exp

ssm_assign, assign to a scheduled variable.

later :: Exp -> Exp -> Exp -> Exp

ssm_later, schedule a delayed assignment to a scheduled variable.

sensitize :: Exp -> Exp -> Exp

ssm_sensitize, sensitize a trigger to a variable.

desensitize :: Exp -> Exp

ssm_sensitize, sensitize a trigger to a variable.

new_adt :: Int -> DConId -> Exp

ssm_new_adt, allocate a new ADT object on the heap.

adt_field :: Exp -> Int -> Exp

ssm_adt_field, access the ith field of an ADT object. Assignable.

adt_tag :: Exp -> Exp

ssm_tag, extract the tag of an ADT value.

adt_heap_tag :: Exp -> Exp

Extract the tag of a heap-allocated ADT value.

container_of :: CIdent

Name of the pseudonymous macro from the Linux kernel.

initialize_program :: CIdent

Name of top level program initialization function

act_typename :: VarId -> CIdent

Obtain the name of a process activation record struct.

act_member :: CIdent

Identifier for act member in act struct.

act_ :: VarId -> Type

Obtain the type of a process activation record.

to_act :: Exp -> VarId -> Exp

Obtain the process-specific activation record from a generic one.

step_ :: VarId -> CIdent

Obtain the name of the step function of a routine.

enter_ :: VarId -> CIdent

Obtain the name for the enter function of a routine.

trig_ :: Int -> CIdent

Obtain the name of each trigger for a routine.

tmp_ :: Int -> CIdent

Obtain the name of a temporary variable.

label_ :: Int -> CIdent

Obtain the name of a label.

arg_ :: Int -> CIdent

Obtain the name of an argument variable.

ret_val :: CIdent

Name of return argument.

actg :: CIdent

Identifier for generic (inner) struct act.

acts :: CIdent

Identifier for specialized (outer) struct act.

acts_ :: CIdent -> Exp

enter_caller :: CIdent

Name of the caller argument of an enter call.

enter_priority :: CIdent

Name of the priority argument of an enter call.

enter_depth :: CIdent

Name of the depth argument of an enter call.

leave_label :: CIdent