Codegen.Typegen

Description

Extract and encapsulate the type information needed during codegen.

Documentation

data TypegenInfo

Type-related information, abstracted behind partial lookup functions.

Constructors

TypegenInfo

Fields

dconInfo :: DConId -> Maybe DConInfo

for each data constructor.

tconInfo :: TConId -> Maybe TConInfo

for each type constructor.

data DConInfo

Information and codegen handlers associated with each data constructor.

Constructors

DConInfo

Fields

dconType :: TConId

the type that the data constructor inhabits

dconSize :: Int

number of fields

dconOnHeap :: Bool

whether the data constructor is heap-allocated

dconCase :: Exp

the dcon tag to match on, i.e., in case tag

dconConstruct :: Exp

constructs a dcon instance

dconDestruct :: Int -> Exp -> Exp

retrieve the ith field

data TConInfo

Information and codegen handlers associated with each type constructor.

Constructors

TConInfo

Fields

typeEncoding :: TypeEncoding

how the data type is encoded

typeScrut :: Exp -> Exp

how to retrieve the tag of an instance

data TypeEncoding

How a data type may be encoded, i.e., heap-allocated, by value, or both.

Constructors

TypePacked

TypeHeap

TypeMixed

genTypes :: [(TConId, TypeDef Type)] -> Pass ([Definition], TypegenInfo)

Create codegen definitions and helpers for sslang type definitions.

genTypeDef :: (TConId, TypeDef Type) -> Pass Definition

Generate C enums for sslang type definitions, enumerating tags.

genTypeInfo :: [(TConId, TypeDef Type)] -> Pass TypegenInfo

Compute codgen helpers for each sslang type definition.