Common.Identifiers

Description

The types of identifiers used across the compiler.

The types VarId, TVarId, DConId, and TConId are used in the IR for data variables, type variables, data constructors, and type constructors, which each inhabit a separate namespace. The underlying Identifier type is used in the AST, before these different kinds of identifiers are differentiated by the lowering phase.

These are defined as newtypes (rather than as type aliases) that ultimately abstract over the underlying String Identifier. We do this for a few reasons:

The particular newtype carries semantic meaning in the IR, e.g., VarId is an identifier for data variables and nothing else. This prevents semantically distinct items from being used in place of one another (without explicit coercion). For example, this will prevent accidentally using a type variable identifier as a data constructor.

In the future, this may be extended to carry other (non-semantic) metadata, such as source code location.

It allows us to freely attach typeclass instances to this type without FlexibleInstances (since String is a type synonym for [Char]).

Users should never need to specifically use the data constructor for each newtype, e.g., VarId (Identifier "foo"). Instead, each identifier belongs to the IsString typeclass, that can be written as fromString "foo", so that the appropriate type can be inferred from the context where the identifier is used. Furthermore, the fromString call can be automatically inserted by the OverloadedStrings GHC extension, meaning we can just write "foo".

All identifier types are instances of the Identifiable typeclass, which allows us to write generic functions that operate over any kind of identifier. One can explicitly convert from an identifier to another using the handy fromId helper; which specific type of identifier may be inferred from the outer context, or explicitly annotated (e.g., fromId i :: VarId to construct a VarId out of identifier i).

Documentation

class (IsString i, Ord i, Show i) => Identifiable i where

A type that may be used as a Sslang identifier.

Methods

ident

Arguments

:: i

-> String
Obtain its underlying String representation.

class IsString a where

Class for string-like datastructures; used by the overloaded string extension (-XOverloadedStrings in GHC).

Methods

fromString :: String -> a

fromId :: (Identifiable a, Identifiable b) => a -> b

Explicitly convert between two types of identifiers.

showId :: (Show a, Identifiable b) => a -> b

Convert a showable instance to some kind of identifier.

newtype TConId

Identifier for type constructors, e.g., Option.

Constructors

TConId Identifier

newtype TVarId

ToIdentifier for type variable, e.g., a.

Constructors

TVarId Identifier

newtype DConId

Identifier for data constructors, e.g., None.

Constructors

DConId Identifier

newtype VarId

Identifier for data variables, e.g., x.

Constructors

VarId Identifier

newtype TVarIdx

de Bruijn index for type variables, e.g., '0

Constructors

TVarIdx Int

newtype CSym

Identifier for C symbols, e.g., printf.

Constructors

CSym Identifier

class Identifiable i => HasFreeVars t i | t -> i where

Terms t that have free variables i

Methods

freeVars :: t -> Set i

newtype Identifier

A generic Sslang identifier.

Used as the type for identifiers in the AST.

Also used as the base type for other identifiers (e.g., TConId, VarId), which derive their typeclass instances from this.

Constructors

Identifier String

isCons :: Identifiable a => a -> Bool

Whether an identifier refers to a type or data constructor.

isVar :: Identifiable a => a -> Bool

Whether an identifier refers to a type or data variable.

Note that internal variables (i.e., isIVar) are also considered variables.

mangle :: (Identifiable i, Data i, Data a) => Proxy i -> a -> a

Mangle all identifiers in some data structure.

This function is useful for preserving the general syntactic structure of a datum without inspecting the specific identifiers used within. This is useful for comparing ASTs modulo alpha renaming.

The Proxy i parameter is used to specify exactly which kind of identifier to mangle. For instance, to mangle all VarId nodes:

mangleVarId :: Data a => a -> a
mangleVarId = mangle (Proxy :: VarId)

mangleVars :: Data a => a -> a

Mangle all type and data variable identifiers.

isReserved :: Identifiable a => a -> Bool

Whether an identifier is reserved.

reserved :: Identifiable a => a -> a

Create a reserved identifier (that can be code-generated as long as a can)

isGenerated :: Identifiable a => a -> Bool

Whether an identifier is an compiler-generated variable name.

genId :: Identifiable a => a -> a

Generate an internal variable name (parenthesized) from some hint.

ungenId :: Identifiable a => a -> Maybe a

Filter out generated identifiers.

cons :: Identifier

Cons identifier for Lists

nil :: Identifier

Nil identifier for Lists