Common.Pretty

Description

Helpers for pretty-printing sslang, including common ASCII tokens.

Documentation

typeAnn :: Pretty t => t -> Doc ann -> Doc ann

typeAnn t d annotates document d with type annotation t.

drarrow :: Doc ann

=>

larrow :: Doc ann

<-

rarrow :: Doc ann

->

dbar :: Doc ann

||

bar :: Doc ann

|

amp :: Doc ann

&

block :: Doc ann -> [Doc ann] -> Doc ann

Constructs a separator-delimited block Doc out of a list of Docs.

spaghetti :: Pretty t => t -> String

Format with a document of infinite width, preventing wraparound.