Front.Scanner

Description

Scanner for sslang.

In addition to tokenizing the input text stream, this scanner is responsible for the following:

Insert implicit open braces after any block-starting token, if necessary;

Insert implicit separators for aligned lines within the same block; and

Insert implicit close braces if an implicit block closed, either by a decrease in indentation, or by the closure of a surrounding explicit block.

As such, the scanner's state is enriched with a stack of contexts which it uses to perform basic delimiter matching. It also relies on epsilon transitions to emit implicit tokens.

Documentation

scanTokens :: String -> Pass [Token]

Extract a token stream from an input string.

scanTokenTypes :: String -> Pass [TokenType]

Extract a stream of token types (without span) from an input string.

lexerForHappy :: (Token -> Alex a) -> Alex a

Used to integrate with Happy parser.

newtype Token

Tokens extracted from source text.

Constructors

Token (Span, TokenType)

data TokenType

The types of tokens that can appear in a sslang source file.

Constructors

TEOF

TType

TIf

TElse

TWhile

TDo

TPar

TLoop

TLet

TMatch

TAfter

TWait

TFun

TExtern

TEq

TLarrow

TRarrow

TDRarrow

TDBar

TColon

TSemicolon

TBar

TComma

TUnderscore

TAt

TAmpersand

TLparen

TRparen

TLbrace

TRbrace

TLbracket

TRbracket

TInteger Integer

TString String

TId Identifier

TOp Identifier

TCSym Identifier

TCQuote String

TCBlock String

data Span

The location of a token in the source text.

Constructors

Span

Fields

tokPos :: Int

tokLen :: Int

tokLine :: Int

tokCol :: Int

syntaxErr :: String -> Alex a

User-facing syntax error.

internalErr :: String -> Alex a

Internal compiler error for unreachable code.

liftErr :: String -> Error

Convert Alex's String-encoded errors to Sslang Error.