Front.Token

Description

Sslang source code tokens.

Documentation

newtype Token

Tokens extracted from source text.

Constructors

Token (Span, TokenType)

tokenType :: Token -> TokenType

Extract the TokenType from a Token.

data Span

The location of a token in the source text.

Constructors

Span

Fields

tokPos :: Int

tokLen :: Int

tokLine :: Int

tokCol :: Int

data TokenType

The types of tokens that can appear in a sslang source file.

Constructors

TEOF

TType

TIf

TElse

TWhile

TDo

TPar

TLoop

TLet

TNow

TAtAt

TBreak

TMatch

TAfter

TWait

TFun

TExtern

TEq

TLarrow

TRarrow

TDRarrow

TDBar

TColon

TSemicolon

TBar

TComma

TUnderscore

TAt

TAmpersand

TLparen

TRparen

TLbrace

TRbrace

TLbracket

TRbracket

TInteger Integer

TString String

TId Identifier

TOp Identifier

TCSym Identifier

TCQuote String

TCBlock String

prettyTokens :: [Token] -> String

Pretty print a list of tokens.