IR

Description

Intermediate representation (IR) stage of the compiler pipeline.

The IR stage is organized into 4 distinct substages: lower, ann2Class, class2Poly, and poly2Flat, delineated by each type system used in the compiler.

Documentation

data Mode

Operation modes for the IR compiler stage.

Constructors

Continue
Compile end-to-end (default).

DumpIR
Print the IR immediately after lowering.

DumpIRTyped
Print the fully-typed IR after type inference.

DumpIRLifted
Print the IR after lambda lifting.

DumpIRFinal
Print the final IR representation before codegen.

data TIType

Constructors

HMOnly
Only run HM type inference

TCOnly
Only run type checker

Both
Run both HM type Inference and type checker

data Options

Compiler options for the IR compiler stage.

Constructors

Options

Fields

mode :: Mode

tiType :: TIType

options :: [OptDescr (Options -> Options)]

CLI options for the IR compiler stage.

lower :: Options -> Program -> Pass (Program Type)

IR compiler sub-stage, lowering AST to optionally type-annotated IR.

ann2Class :: Options -> Program Type -> Pass (Program Type)

IR compiler sub-stage, ultimately producing a fully-typed IR.

Runs HM inference, type checker, or both (default) based on the CLI option.

When both algorithms are run, the program fails type check iff both algorithms throw an error.

class2Poly :: Options -> Program Type -> Pass (Program Type)

IR compiler sub-stage, ultimately instantiating all type classes.

poly2Poly :: Options -> Program Type -> Pass (Program Type)

IR compiler sub-stage, performing source-to-source translations.

run :: Options -> Program -> Pass (Program Type)

IR compiler stage.

setHM :: Options -> Options

Helper function to set ti type to HM-only

setTC :: Options -> Options

Helper function to set ti type to TC-only