IR.Constraint.Canonical

Documentation

data Type

Encoding of sslang types.

Structurally speaking, these are very simple. Types are either type variables or type constructors applied to some other types.

Builtin types (and type constructors) include Arrow, Unit, Ref, List, and various sizes of tuples; for convenience, those are defined elsewhere using the GHC PatternSynonyms extension.

Constructors

TCon TConId [Type]
Type constructor, applied to zero or more types

TVar TVarId
Type variable (may be implicitly quantified)

data Scheme

SCHEMES

Constructors

Forall FreeVars Type

type FreeVars = Map TVarId ()

data Annotation

An annotation records the annotated portion of a pattern.

Constructors

AnnType Type
A basic Type annotation

AnnDCon DConId [Annotation]
Annotations collected from patterns

AnnArrows [Annotation] Annotation
Annotations collected from fun args

newtype Annotations

Expressions are annotated with a (potentially empty) list of Annotation.

Constructors

Annotations [Annotation]

type Kind = Int

KINDS

schemeOf :: Type -> Scheme

Construct a scheme from all free type variables and a trivial constraint.

freeVars :: Type -> FreeVars

foldArrow :: ([Type], Type) -> Type

Fold a list of argument types and a return type into an Arrow Type.

unfoldArrow :: Type -> ([Type], Type)

Unfold an Arrow Type into a list of argument types and a return type.

(-->) :: Type -> Type -> Type infixr 0

HELPERS

pattern Unit :: Type

The builtin singleton Type, whose only data constructor is just ().

pattern Ref :: Type -> Type

The builtin reference Type, created using new.

pattern List :: Type -> Type

The builtin list Type, created using list syntax, e.g., [a, b].

pattern Time :: Type

The builtin 64-bit timestamp Type.

pattern I64 :: Type

Builtin Type for signed 64-bit integers.

pattern U64 :: Type

Builtin Type for unsigned 64-bit integers.

pattern I32 :: Type

Builtin Type for signed 32-bit integers.

pattern U32 :: Type

Builtin Type for unsigned 32-bit integers.

pattern I16 :: Type

Builtin Type for signed 16-bit integers.

pattern U16 :: Type

Builtin Type for unsigned 16-bit integers.

pattern I8 :: Type

Builtin Type for signed 8-bit integers.

pattern U8 :: Type

Builtin Type for unsigned 8-bit integers.

tuple :: [Type] -> Type

Construct a builtin tuple type out of a list of at least 2 types.

unAnnotations :: Annotations -> [Annotation]

Unwrap the Annotations data constructor.

annToType :: Annotation -> Type

ANNOTATION

builtinKinds :: Map TConId Kind

Map to help us look up the kinds of builtin types.