IR.Constraint.Solve

Documentation

run :: Constraint -> TC ()

RUN SOLVER

emptyState :: State

type Env = Map Identifier Variable

SOLVER

type Pools = IOVector [Variable]

data State

Constructors

State

Fields

_env :: Env

_mark :: Mark

_errors :: [Error]

solve :: Env -> Int -> Pools -> State -> Constraint -> TC State

isGeneric :: Variable -> TC ()

addError :: State -> Error -> State

ERROR HELPERS

occurs :: State -> (Identifier, Variable) -> TC State

OCCURS CHECK

generalize :: Mark -> Mark -> Int -> Pools -> TC ()

GENERALIZE

Every variable has rank less than or equal to the maxRank of the pool. This sorts variables into the young and old pools accordingly.

poolToRankTable :: Mark -> Int -> [Variable] -> TC (Vector [Variable])

adjustRank :: Mark -> Mark -> Int -> Variable -> TC Int

ADJUST RANK

adjustRankContent :: Mark -> Mark -> Int -> Content -> TC Int

introduce :: Int -> Pools -> [Variable] -> TC ()

REGISTER VARIABLES

typeToVariable :: Int -> Pools -> Type -> TC Variable

TYPE TO VARIABLE

typeToVar :: Int -> Pools -> Map TVarId Variable -> Type -> TC Variable

register :: Int -> Pools -> Content -> TC Variable

schemeToVariable :: Int -> Pools -> Map TVarId () -> Type -> TC Variable

schemeToVar :: Int -> Pools -> Map TVarId Variable -> Type -> TC Variable

makeCopy :: Int -> Pools -> Variable -> TC Variable

COPY

makeCopyHelp :: Int -> Pools -> Variable -> TC Variable

restore :: Variable -> TC ()

restoreContent :: Content -> TC ()

traverseFlatType :: (Variable -> TC Variable) -> FlatType -> TC FlatType