IR.HM

Description

Infer types for IR of optionally annotated program using Hindley Milner type inference algorithm with union find tree data structure.

There are three main stages of the algorithm 1. Assign symbolic typenames During this stage, we walk through each expression recursively, initialize its type by either assigning a fresh type variable or looking it up in the type context at that point, and generate a list of type equations to record the relationships between types. We also collapse the list of type annotations into the most specific one and put that information into a type equation as well to bring it into the later inference. Note that we will be modifying the expression monad simultaneously to record its new type info. 2. Solve type equations using unification. During this stage, we solve the type equations generated in the last stage and build a union find tree. This stage should be merged with stage 1 in the future. 3. Find the type of the expression. During this stage, we walk through each expression recursively and assign it with its inferred type by loop up the root of its temporary type in the union find tree.

The three stages are encaptured by the functions initTypeVars, unifyAll, and getType.

TODO: - Define a partial order for the generality of types. The order will be used to (1) collapse type annotations into the most specific one; (2) check that type annotation is equal to or more specific than the expression's inferred type. - Handle recursive function correctly.

Documentation

data InferState

Inference State.

Constructors

InferState

Fields

varMap :: Map VarId Scheme

equations :: [(Type, Type)]

unionFindTree :: Map Type Type

count :: Int

dConType :: Map DConId (TConId, [TVarId], [Type])

newtype InferFn a

Inference Monad.

Constructors

InferFn (StateT InferState Pass a)

runInferFn :: InferFn a -> Pass a

Run a InferFn computation.

inferProgram :: Program Type -> Pass (Program Type)

inferProgram p infers the type of all the programDefs of the given program p.

recordFDefs :: [(VarId, Expr Type)] -> InferFn ()

recordFDefs saves information about all the declared functions for future use

recordADTs :: [(TConId, TypeDef Type)] -> InferFn [(TConId, TypeDef Type)]

recordADTs saves information about ADTs in the Inference State for future use

saves type of DCon as a (DCon,TCon) key-value pair transforms Ann.Type to Classes.Type saves types of DCon args (the ADT fields) as (DCon, [type]) key-value pairs

recordDCon :: TConId -> [TVarId] -> (DConId, TypeVariant Type) -> InferFn ()

insertType inserts the overall type of an ADT's DConid into the Inference State

recordExterns :: [(VarId, Type)] -> InferFn [(VarId, Type)]

Record the types of extern symbols. NOTE: this does not actually check for arity and probably doesn't do polymorphism correctly either.

inferProgramDefs :: [(VarId, Expr Type)] -> InferFn [(VarId, Expr Type)]

inferProgramDefs ds infers the type of programDefs ds recursively and binds each varibale to its type.

initAllTypeVars :: [(VarId, Expr Type)] -> InferFn [(VarId, Expr Type)]

initAllTypeVars initialize all the funcitons' type at once

getAllTypes :: [(VarId, Expr Type)] -> InferFn [(VarId, Expr Type)]

getAllTypes replace all the type variables in the initial funciton types

class HasFreeTVars a where

Class of things that may contain free type variables.

Methods

freeTVars :: a -> Set TVarId

instantiate :: Scheme -> InferFn Type

Instantiates the type from the scheme Forall vs t.

Create fresh type variable for each quantified type variable in vs and substitute them in t by applying subst to t.

generalize :: Type -> InferFn Scheme

Generalizes the type t into a type scheme given the typing context ctx.

Collect all the free type variables in the type t and all the free type variables in the typing context ctx, away any type variables that appear free in the type environment, and whatever is leftover can be generalized into a type scheme.

letters :: [String]

letters represents list of identifiers to be used to construct free type variables assigned during type inference. The leading '_' distinguish them from user-annotated type variables.

dummyAnnT :: Type

dummyAnnT represents empty type annotation list.

fresh :: InferFn Type

Generates a new TVar using the next tick number and increment the counter.

lookupVar :: VarId -> InferFn (Maybe Scheme)

lookupVar v looks up the type scheme of a variable in the current varMap given its variable ID v.

lookupDCon :: DConId -> InferFn (Maybe Scheme)

Look up the type of a data constructor in the inference state.

insertVar :: VarId -> Scheme -> InferFn ()

insertVar v t Insert a variable ID and its type scheme into current varMap.

insertBinder :: Binder -> Scheme -> InferFn ()

insertBinder is a wrapper of insertVar for Binder type.

insertEquation :: (Type, Type) -> InferFn ()

insertEquation e inserts e into the current equations list.

insertUnion :: Type -> Type -> InferFn ()

insertUnion t1 t2 inserts t1 : t2 into the current unionFindTree.

withNewScope :: InferFn a -> InferFn a

Helper function to support local modificaton of varMap.

withNewTypeScope :: InferFn a -> InferFn a

Helper function to support local modificaton of unionFindTree.

collapseAnnT :: Type -> Type

Helper function to collapse a list of type annotation to a single one with 'Classes.type' type. If the list is empty, use Classes.TVar (TVarId $ Identifier "_")) to denote a dummy annotation. TODO: Currently just take the first type annotation of the list as the result. Need to implement the collapse step to find and return the most specific type annotation from the list. Need yo add support for annotating type variables as well.

initTypeVars :: Expr Type -> InferFn (Expr Type)

Stage 1: Assign symbolic typenames

initTypeVars e walks into the expression e recursively, assign a fresh TVar to each unknown type, and build type equations that will be solved later.

inferPrimBinop :: PrimOp -> Expr Type -> Expr Type -> Type -> InferFn (Expr Type)

Infer the type of a binary integer operation.

typeCheck :: Type -> Type -> InferFn Type

typeCheck annT expectedT checks whether the type annotation annT is compatible with the expected type expectedT. Insert new type equations when the expected type is a type variable. TODO: currently only consider equality as compatibility. Need to support the case when type annotation is more specific than the expected type.

unifyAll :: InferFn ()

Stage 2: Solve type equations using unification.

Solve the list of equations one by one and puts solutions as new enties in the unionFindTree.

unify :: Type -> Type -> InferFn ()

unify t1 t2 solves the type equation t1 ~ t2 and put the solution into unionFindTree.

getType :: Expr Type -> InferFn (Expr Type)

Stage 3: Find the type of the expression based on unionFindTree.

getType e walks into the expression e and solves its type t recursively.

getTypePrimOp :: PrimOp -> [Expr Type] -> Type -> InferFn (Expr Type)

Recursively get type of operands to PrimOp, then apply given type.

solveType :: Type -> InferFn Type

solveType t solves the type t by replacing any embeded TVar tvar with its real type, which is the root of tvar in the unionFindTree.

findRoot :: Type -> InferFn Type

findRoot t finds the root of t inside unionFindTree.