IR.IR

Description

Sslang's intermediate representation and its associated helpers.

Documentation

data Program t

Top-level compilation unit.

t is the type system in use, e.g., IR.Types.Flat

Constructors

Program

Fields

programEntry :: VarId

cDefs :: String

externDecls :: [(VarId, Type)]

programDefs :: [(VarId, Expr t)]

typeDefs :: [(TConId, TypeDef)]

data TypeDef

The type definition associated with a type constructor. A definition for `data MyList a = Cons a (MyList a) | Nil` looks like: TypeDef { targs = [a] , [ (Cons, VariantUnnamed [TVar a, TCon (MyList [TVar a])]) , (Nil, VariantUnnamed []) ] } (Data constructors for identifiers are omitted for brevity.) Note that for a flat type system, where all type constructors are nullary, targs will just be set to [].

Constructors

TypeDef

Fields

variants :: [(DConId, TypeVariant)]

targs :: [TVarId]

data TypeVariant

Arguments to a data constructor, whose fields may or may not be named

Constructors

VariantNamed [(VarId, Type)]
A record with named fields

VariantUnnamed [Type]
An algebraic type with unnamed fields

type Binder = Maybe VarId

A name to be bound; Nothing represents a wildcard, e.g., let _ = ....

data Literal

Literal values supported by the language.

Note that these don't carry any connotation of type: 1 just means 1,

Constructors

LitIntegral Integer

LitEvent

data PrimOp

Primitive operations.

These should be the kinds of functions one may expect to be available as operators in C, or as instructions in an assembly language.

For simplicity and consistency, they should be:

Pure (i.e., side-effectful iff operands are side-effectful, i.e., no =)

Strict in all operands (i.e., no && or ||)

We can instead implement short-circuit control flow using match statements.

Constructors

PrimNeg
negation, i.e., -x

PrimNot
logical not, i.e., !x

PrimBitNot
bitwise not, i.e., ~x

PrimAdd
addition, i.e., x + y

PrimSub
subtraction, i.e., x - y

PrimMul
multiplication, i.e., x * y

PrimDiv
division, i.e., x / y

PrimMod
modulus, i.e., x % y

PrimBitAnd
bitwise-and, i.e., x & y

PrimBitOr
bitwise-or, i.e., x | y

PrimEq
equality, i.e., x == y

PrimNeq
equality, i.e., x != y

PrimGt
greater than, i.e., x > y

PrimGe
greater than or equal to, i.e., x >= y

PrimLt
less than, i.e., x < y

PrimLe
less than or equal to, i.e., x <= y

data Primitive

Primitive functions for side-effects and imperative control flow.

Constructors

New
New e allocates a schedule variable initialized to e, and returns a reference to it.

Dup
Dup r dups the reference r and returns r.

Drop
Drop e r evaluates to e, but also drops r.

Deref
Deref r dereferences reference r to obtain its value.

Assign
Assign r e instantly assigns value e to reference r.

After
After t r e assigns e to reference r after time t.

Par
Par es+ evaluates expressions es concurrently.

Wait
Wait rs+ waits for an assignment to any reference in rs.

Loop
Loop b loops body b forever.

Break
Break breaks out of the innermost loop.

Now
Now obtains the value of the current instant

PrimOp PrimOp
Inlined C expression code.

CQuote String
Primitive operator.

CCall CSym
Direct call to arbitrary C function (NOTE: HACKY).

FfiCall VarId
Call to well-typed extern symbol.

data Expr t

Expressions, based on the let-polymorphic lambda calculus.

t represents the type of this expression, e.g., Flat. At various stages, this may represent a richer or simpler type system. The type is embedded in each data constructor so as to type-annotate the entire expression tree.

Designed for side effects with call-by-value evaluation order. Basic sequencing can be recovered through let-bindings:

let _ = stmt1 in
let _ = stmt2 in
...

Effects of stmt1 take place before that of stmt2.

Constructors

Var VarId t
Var n t is a variable named n of type t.

Data DConId t
Data d t is a data constructor named d of type t.

Lit Literal t
Lit l t is a literal value l of type t.

App (Expr t) (Expr t) t
App f a t applies function f to argument a, producing a value of type t.

Let [(Binder, Expr t)] (Expr t) t
Let [(n, v)] b t binds value v to variable v in its body b.

The bindings list may only be of length greater than 1 for a set of mutually co-recursive functions.

Lambda Binder (Expr t) t
Lambda v b t constructs an anonymous function of type t that binds a value to parameter v in its body b.

Match (Expr t) [(Alt, Expr t)] t
Match s alts t pattern-matches on scrutinee s against alternatives alts, each producing a value of type t.

Prim Primitive [Expr t] t
Prim p es t applies primitive p arguments es, producing a value of type t.

Exception ExceptType t
Exception et t produces a exception for the program.

data Alt

An alternative in a pattern-match.

Constructors

AltData DConId [Alt]
AltData d vs matches data constructor d, and recursive patterns alts.

AltLit Literal
AltLit l matches against literal l, producing expression e.

AltBinder Binder
AltBinder v matches anything, and bound to name v.

newtype VarId

Identifier for data variables, e.g., x.

Constructors

VarId Identifier

newtype TConId

Identifier for type constructors, e.g., Option.

Constructors

TConId Identifier

newtype DConId

Identifier for data constructors, e.g., None.

Constructors

DConId Identifier

newtype ExceptType

Constructors

ExceptDefault Literal

data Type

Encoding of sslang types.

Structurally speaking, these are very simple. Types are either type variables or type constructors applied to some other types.

Builtin types (and type constructors) include Arrow, Unit, Ref, List, and various sizes of tuples; for convenience, those are defined elsewhere using the GHC PatternSynonyms extension.

data Annotation

An annotation records the annotated portion of a pattern.

data Annotations

Expressions are annotated with a (potentially empty) list of Annotation.

variantFields :: TypeVariant -> Int

The number of fields in a TypeVariant.

wellFormed :: Expr t -> Bool

Predicate of whether an expression "looks about right".

Description left deliberately vague so that we have the flexibility to strengthen this predicate. For now, we just check that all primitives are applied to the right number of arguments.

This predicate also provides a template to recursively traverse through all sub-expressions of an expression.

foldLambda :: [(Binder, Type)] -> Expr Type -> Expr Type

Create a lambda chain given a list of argument-type pairs and a body.

unfoldLambda :: Expr t -> ([Binder], Expr t)

Collect a curried list of function arguments from a nesting of lambdas.

extract :: Expr t -> t

Extract the type carried by an Expr.

inject :: t -> Expr t -> Expr t

Replace the top-level type carried by an Expr.

foldApp :: Expr t -> [(Expr t, t)] -> Expr t

Apply a function to zero or more arguments.

foldApp is the inverse of unfoldApp.

unfoldApp :: Expr t -> (Expr t, [(Expr t, t)])

Collect a curried application into the function and argument list.

The type accompanying each argument represents type produced by the application, and is extracted from the App node that this function unwraps.

For example, the term f a b (where a: A and b: B) would be represented by the following AST: @ (App (App (Var f (A -> B -> C)) (Var a A) (B -> C)) (Var b B) C) @

which, when unzipped, gives:

@ (Var f (A -> B -> C)) [(Var a A, B -> C), (Var b B, C)] @

unfoldApp is the inverse of foldApp.

isValue :: Expr t -> Bool

Whether an expression is a value.

getAltDefault :: Alt -> Binder

Retrieve binder from Alt, assuming it is AltBinder.