IR.LowerAst

Description

Lower the representation of a sslang Ast into sslang IR.

This pass expects prior desugaring passes to ensure that:

Op regions are unflatted into proper applications.

Patterns in definitions consist of only (annotated) identifiers or wildcards.

Documentation

lowerProgram :: Program -> Pass (Program Annotations)

Lower an AST Program into IR.