IR.MangleNames

Description

Mangle all variable names in a program to ensure global uniqueness.

Also populate symbol table with information about those (now unique) names.

This module also exports pickId, which other passes can use to create fresh variable names given some symbol table.

Documentation

mangleProgram :: Program Type -> Pass (Program Type)

normalizeId :: Identifiable i => i -> i

Normalize an identifier (remove things like backticks etc.)

pickId :: Map MangledId t -> OriginId -> MangledId

Given a symbol table and a name, pick a new name that is globally unique.