IR.Pretty

Orphan instances

Pretty ExceptType

Instance details

Methods

pretty :: ExceptType -> Doc ann

prettyList :: [ExceptType] -> Doc ann

Pretty PrimOp

Instance details

Methods

pretty :: PrimOp -> Doc ann

prettyList :: [PrimOp] -> Doc ann

Pretty Literal

Instance details

Methods

pretty :: Literal -> Doc ann

prettyList :: [Literal] -> Doc ann

Pretty (Alt Type)

Instance details

Methods

pretty :: Alt Type -> Doc ann

prettyList :: [Alt Type] -> Doc ann

Pretty (Expr Type)

Instance details

Methods

pretty :: Expr Type -> Doc ann

prettyList :: [Expr Type] -> Doc ann

Pretty (Program Type)
Pretty Typeclass: pretty print the IR

Adds * indentation and line breaks * some parens (not minimal parens, but fewer than around every node) Omits * let _ = * type annotations Reverts * curried funcs of one arg back to multiple arg funcs

Instance details

Methods

pretty :: Program Type -> Doc ann

prettyList :: [Program Type] -> Doc ann