IR.TypeChecker

Description

Infer types for optionally annotated program.

For now, this module implements a very crude form of pure, syntax-directed type inference where type annotations can be locally and immediately inferred, without the help of a type environment/context.

Documentation

data TypeCtx

Typing Environment

Constructors

TypeCtx

Fields

varMap :: Map VarId Type

dConMap :: Map DConId Type

newtype InferFn a

Constructors

InferFn (StateT TypeCtx Pass a)

runInferFn :: InferFn a -> Pass a

Run a InferFn computation.

insertVar :: VarId -> Type -> InferFn ()

Insert a variable ID and its type into current typing context.

replaceCtx :: TypeCtx -> InferFn ()

Replace the current typing context with a new one.

lookupVar :: VarId -> InferFn (Maybe Type)

Look up the type of a variable in the current typing context given its variable ID.

lookupDCon :: DConId -> InferFn (Maybe Type)

Look up the type of a data constructor in the typing context.

inferProgram :: Program Type -> Pass (Program Type)

Infer all the program defs of the given program.

inferADT :: [(TConId, TypeDef Type)] -> InferFn [(TConId, TypeDef Type)]

Pass all program typeDefs through the typechecker

Change Ann.Type to Classes.Type For each ADT, save ('DConId, 'TConId) key-value pairs in typing context for future use

inferExterns :: [(VarId, Type)] -> InferFn [(VarId, Type)]

inferProgramDefs :: [(VarId, Expr Type)] -> InferFn [(VarId, Expr Type)]

Top level inference.

This will infer the (VarId, Expr) pair sequentially and add the infered types into the type context so that these info can be used when infering the following expressions.

inferExpr :: Expr Type -> InferFn (Expr Type)

Infer a single expression e.

This function will also infer any nested expressions inside e and all expressions in the return value will have Classes.Type as its type. Error will be thrown if the type annotation is in inconsistent with the true true or the expression is known.

inferPrim :: Expr Type -> InferFn (Expr Type)

Infer the type of a Primitive expression.

inferPrimBinop :: PrimOp -> Expr Type -> Expr Type -> InferFn (Expr Type)

Check type of a binary integer operation.

withNewScope :: InferFn a -> InferFn a

Helper function to support local modificaton of type context.

withVty :: Expr Type -> Binder -> Type -> InferFn ()

Helper function to do type inference with the knowledge of the correct type for a binding.

anns2Class :: Type -> InferFn Type

Transfrom an Ann.Type to Classes.Type.

ann2Class :: TypeAnnote -> InferFn Type