IR.Types.Annotated

Description

Type annotations, as collected from the Ast.

Documentation

data Builtin t

Builtin type constructors that should be preserved across all type systems.

Constructors

Unit
Singleton type ()

Void
Uninhabited type !, coerces with anything

Ref t
Reference of &

Arrow t t
Function arrow a -> b

Tuple [t]
Tuple with two or more fields

Integral Int
Two's complement binary type with size in bits

newtype Type

A single term may be annotated by zero or more types.

When multiple exist, it should be assumed that they are equivalent, in the sense that they can be unified.

Type annotations can be added using <> (from Semigroup), while mempty represents no type annotation.

Constructors

Type [TypeAnnote]

data TypeAnnote

A type annotation.

Constructors

TBuiltin (Builtin Type)
Builtin types

TCon TConId [Type]
Type constructor, e.g., Option '0

TVar TVarId
Type variables, e.g., '0

untyped :: Type

Convenience helper for no type annotations.