IR.Types.Type

Description

Definitions of and related to the sslang IR's type system.

Documentation

data Type

Encoding of sslang types.

Structurally speaking, these are very simple. Types are either type variables or type constructors applied to some other types.

Builtin types (and type constructors) include Arrow, Unit, Ref, List, and various sizes of tuples; for convenience, those are defined elsewhere using the GHC PatternSynonyms extension.

Constructors

TCon TConId [Type]
Type constructor, applied to zero or more types

TVar TVarId
Type variable (may be implicitly quantified)

data Constraint

Constraints on a type scheme.

For now, we only support trivial constraints.

Constructors

CTrue
The trivial constraint, i.e., always satisfied.

data SchemeOf t

Schemes quantify over Type variables and impose some Constraint.

SchemeOf is implemented as functor over some kind of type so we can easily substitute in Type vs UType when performing type inference/unification.

Constructors

Forall (Set TVarId) Constraint t

unScheme :: SchemeOf t -> t

Unwrap a scheme and obtain the underlying type.

forall :: (Functor l, Foldable l) => l TVarId -> t -> SchemeOf t

Construct a scheme with quantified type variables and a trivial constraint.

schemeOf :: Type -> Scheme

Construct a scheme from all free type variables and a trivial constraint.

newtype Scheme

Schemes over Type.

Constructors

Scheme (SchemeOf Type)

data Annotation

An annotation records the annotated portion of a pattern.

Constructors

AnnType Type
A basic Type annotation

AnnDCon DConId [Annotation]
Annotations collected from patterns

AnnArrows [Annotation] Annotation
Annotations collected from fun args

newtype Annotations

Expressions are annotated with a (potentially empty) list of Annotation.

Constructors

Annotations [Annotation]

unAnnotations :: Annotations -> [Annotation]

Unwrap the Annotations data constructor.

fromAnnotations :: Annotations -> Type

Unroll an annotation into a type. FIXME: get rid of this.

class HasType a where

Some data type that contains a sslang Type.

Methods

getType :: a -> Type

pattern Hole :: Type

A fresh, free type variable; may appear in type annotations.

pattern Arrow :: Type -> Type -> Type

The type constructor for function arrows.

unfoldArrow :: Type -> ([Type], Type)

Unfold an Arrow Type into a list of argument types and a return type.

foldArrow :: ([Type], Type) -> Type

Fold a list of argument types and a return type into an Arrow Type.

pattern Unit :: Type

The builtin singleton Type, whose only data constructor is just ().

pattern Ref :: Type -> Type

The builtin reference Type, created using new.

pattern List :: Type -> Type

The builtin list Type, created using list syntax, e.g., [a, b].

pattern Time :: Type

The builtin 64-bit timestamp Type.

pattern I64 :: Type

Builtin Type for signed 64-bit integers.

pattern U64 :: Type

Builtin Type for unsigned 64-bit integers.

pattern I32 :: Type

Builtin Type for signed 32-bit integers.

pattern U32 :: Type

Builtin Type for unsigned 32-bit integers.

pattern I16 :: Type

Builtin Type for signed 16-bit integers.

pattern U16 :: Type

Builtin Type for unsigned 16-bit integers.

pattern I8 :: Type

Builtin Type for signed 8-bit integers.

pattern U8 :: Type

Builtin Type for unsigned 8-bit integers.

isInt :: Type -> Bool

Test whether a Type is one of the builtin signed integers.

isUInt :: Type -> Bool

Test whether a Type is one of the builtin unsigned integers.

isNum :: Type -> Bool

Test whether a Type is one of the builtin numeric types.

tuple :: [Type] -> Type

Construct a builtin tuple type out of a list of at least 2 types.

isTuple :: Type -> Bool

Test whether a Type is a tuple of some arity.

tupleId :: (Integral i, Identifiable v) => i -> v

Construct the type constructor of a builtin tuple of given arity (>= 2).

data TupleView

More convenient representation of tuple types, for pattern-matching.

Constructors

Tup2 (Type, Type)
2-tuples

Tup3 (Type, Type, Type)
3-tuples

Tup4 (Type, Type, Type, Type)
4-tuples

TupN [Type]
n-ary tuples

NotATuple
not a tuple

tupleOf :: Type -> TupleView

Convert a Type to a TupleView; convenient for ViewPatterns.

For example, to match on just 2-tuples;

foo :: Type -> String
foo (tupleOf -> Tup2 (a, b)) = "2-tuple of " ++ show a ++ " and " ++ show b
foo t                        = "Some other kind of type: " ++ show t

type Kind = Int

Kinds are just the arity of type constructors.

builtinKinds :: Map TConId Kind

Map to help us look up the kinds of builtin types.