IR.Types.TypeSystem

Description

Common definitions for the type system(s) used in this compiler.

The following is a sketch of the type pipeline (from top to bottom):

Types.Ast: type classes + polymorphic types + implicit types (to be inferred)

(type inference)

Types.Classes: type classes + polymorphic types

(typeclass instantiation)

Types.Poly: polymorphic types

(monomophisation)

Types.Flat: concrete only

Documentation

type Arity = Int

The number of arguments a type constructor will take.

data Builtin t

Builtin type constructors that should be preserved across all type systems.

Constructors

Unit
Singleton type ()

Void
Uninhabited type !, coerces with anything

Ref t
Reference of &

Arrow t t
Function arrow a -> b

Tuple [t]
Tuple with two or more fields

Integral Int
Two's complement binary type with size in bits

class TypeSystem t where

A type system must allow us to construct and access underlying builtins.

Instances should satisfy the following law:

injectBuiltin . projectBuiltin = Just

Methods

projectBuiltin :: Builtin t -> t

injectBuiltin :: t -> Maybe (Builtin t)

unit :: TypeSystem t => t

Helper to construct Unit type in any TypeSystem.

void :: TypeSystem t => t

Helper to construct Void type in any TypeSystem.

ref :: TypeSystem t => t -> t

Helper to construct Ref type in any TypeSystem.

arrow :: TypeSystem t => t -> t -> t

Helper to construct Arrow type in any TypeSystem.

deref :: TypeSystem t => t -> Maybe t

Helper to unwrap Ref in any TypeSystem.

dearrow :: TypeSystem t => t -> Maybe (t, t)

Helper to unwrap Arrow in any TypeSystem.

tuple :: TypeSystem t => [t] -> t

Helper to construct tuples from list of types, accounting for size < 2.

int :: TypeSystem t => Int -> t

Helper to construct integral type.

collectArrow :: TypeSystem t => t -> ([t], t)

Decompose an Arrow type into a list of argument types and a return type.

data TypeDef t

The type definition associated with a type constructor.

A definition for `data MyList a = Cons a (MyList a) | Nil` looks like:

  TypeDef { targs = [a]
          , [ (Cons, VariantUnnamed [TVar a, TCon (MyList [TVar a])])
            , (Nil, VariantUnnamed [])
            ]
          }

(Data constructors for identifiers are omitted for brevity.)

Note that for a flat type system, where all type constructors are nullary, targs will just be set to [].

Constructors

TypeDef

Fields

variants :: [(DConId, TypeVariant t)]

targs :: [TVarId]

data TypeVariant t

Arguments to a data constructor, whose fields may or may not be named

Constructors

VariantNamed [(VarId, t)]
A record with named fields

VariantUnnamed [t]
An algebraic type with unnamed fields

variantFields :: TypeVariant t -> Int

The number of fields in a TypeVariant.