IR.Types.Unification

Description

Wrapper around the unification-fd library.

Compared to normal types (Type), unification types (Type) may contain unification variables, which only appear during the type inference process.

This module provides a definition for unification types, an inference monad, as well as some pattern synonyms for the builtin types.

Unification types

type Type = UTerm TypeF IntVar

A type that may also contain unification variables.

type Scheme = SchemeOf Type

An explicitly quantified Type, which may contain unification variables.

pattern TVar :: TVarId -> Type

Pattern synonym for Type variables.

pattern TCon :: TConId -> [Type] -> Type

Pattern synonym for Type constructors.

class HasFreeUVars a where

Methods

freeUVars :: a -> InferM ctx (Set IntVar)

The inference monad

type InferM ctx = ReaderT ctx (ExceptT Error (IntBindingT TypeF Pass))

Inference monad, build on top of the unification algorithm.

runInfer :: ctx -> InferM ctx a -> Pass a

Execute the InferM inference monad.

fresh :: InferM ctx Type

Create a fresh unification variable.

applyBindings :: Type -> InferM ctx Type

Apply bindings from current monad so remaining bindings must be free.

(=:=) :: Type -> Type -> InferM ctx () infix 4

Unifies two types; fails if they cannot unify.

(<:=) :: Type -> Type -> InferM ctx Bool infix 4

Whether the LHS subsumes the RHS.

unfreeze :: Type -> Type

Promote a Type to a Type.

freeze :: Type -> InferM ctx Type

Demote a Type to a regular Type.

Fails if the unification type contains more than just contructors and variables.

instantiate :: Scheme -> InferM ctx Type

Instantiate a Scheme by replacing all quantified type varibles with fresh unification variables

generalize :: HasFreeUVars ctx => Type -> InferM ctx Scheme

Generalize a Type by replacing all free unification variables with quantified type variables.

class Monad m => MonadReader r (m :: Type -> Type) | m -> r where

See examples in Control.Monad.Reader. Note, the partially applied function type (->) r is a simple reader monad. See the instance declaration below.

Minimal complete definition

(ask | reader), local

Methods

ask :: m r

Retrieves the monad environment.

local

Arguments

:: (r -> r)
The function to modify the environment.

-> m a
Reader to run in the modified environment.

-> m a

Executes a computation in a modified environment.

reader

Arguments

:: (r -> a)
The selector function to apply to the environment.

-> m a

Retrieves a function of the current environment.

class Monad m => MonadError e (m :: Type -> Type) | m -> e where

The strategy of combining computations that can throw exceptions by bypassing bound functions from the point an exception is thrown to the point that it is handled.

Is parameterized over the type of error information and the monad type constructor. It is common to use Either String as the monad type constructor for an error monad in which error descriptions take the form of strings. In that case and many other common cases the resulting monad is already defined as an instance of the MonadError class. You can also define your own error type and/or use a monad type constructor other than Either String or Either IOError. In these cases you will have to explicitly define instances of the MonadError class. (If you are using the deprecated Control.Monad.Error or Control.Monad.Trans.Error, you may also have to define an Error instance.)

Methods

throwError :: e -> m a

Is used within a monadic computation to begin exception processing.

catchError :: m a -> (e -> m a) -> m a

A handler function to handle previous errors and return to normal execution. A common idiom is:

do { action1; action2; action3 } `catchError` handler

where the action functions can call throwError. Note that handler and the do-block must have the same return type.

asks

Arguments

:: MonadReader r m

=> (r -> a)
The selector function to apply to the environment.

-> m a

Retrieves a function of the current environment.

Builtin types

pattern Hole :: Type

A fresh, free type variable; may appear in type annotations.

pattern Arrow :: Type -> Type -> Type

The type constructor for function arrows.

foldArrow :: ([Type], Type) -> Type

Fold a list of argument types and a return type into an Arrow Type.

unfoldArrow :: Type -> ([Type], Type)

Unfold an Arrow Type into a list of argument types and a return type.

pattern Unit :: Type

The builtin singleton Type, whose only data constructor is just ().

pattern Ref :: Type -> Type

The builtin reference Type, created using new.

pattern List :: Type -> Type

The builtin list Type, created using list syntax, e.g., [a, b].

pattern Time :: Type

The builtin 64-bit timestamp Type.

pattern I64 :: Type

Builtin Type for signed 64-bit integers.

pattern U64 :: Type

Builtin Type for unsigned 64-bit integers.

pattern I32 :: Type

Builtin Type for signed 32-bit integers.

pattern U32 :: Type

Builtin Type for unsigned 32-bit integers.

pattern I16 :: Type

Builtin Type for signed 16-bit integers.

pattern U16 :: Type

Builtin Type for unsigned 16-bit integers.

pattern I8 :: Type

Builtin Type for signed 8-bit integers.

pattern U8 :: Type

Builtin Type for unsigned 8-bit integers.

tuple :: [Type] -> Type

Construct a builtin tuple type out of a list of at least 2 types.

Orphan instances

Ord IntVar

Instance details

Methods

compare :: IntVar -> IntVar -> Ordering

(<) :: IntVar -> IntVar -> Bool

(<=) :: IntVar -> IntVar -> Bool

(>) :: IntVar -> IntVar -> Bool

(>=) :: IntVar -> IntVar -> Bool

max :: IntVar -> IntVar -> IntVar

min :: IntVar -> IntVar -> IntVar

Unifiable []

Instance details

Methods

zipMatch :: [a] -> [a] -> Maybe [Either a (a, a)]