Common.Pretty

Description

Helpers for pretty-printing sslang, including common ASCII tokens.

Documentation

typeAnn :: Dumpy t => t -> Doc ann -> Doc ann

typeAnn t d annotates document d with type annotation t.

Only used by spaghetti so constrained by Dumpy instead of Pretty for now.

drarrow :: Doc ann

=>

larrow :: Doc ann

<-

rarrow :: Doc ann

->

dbar :: Doc ann

||

bar :: Doc ann

|

amp :: Doc ann

&

block :: Doc ann -> [Doc ann] -> Doc ann

Constructs a separator-delimited block Doc out of a list of Docs.

indentNo :: Int

Number of spaces to indent

spaghetti :: Dumpy t => t -> String

Format with a document of infinite width, preventing wraparound.

class Dumpy a where

Lengthy Typeclass: print an ugly but parseable representation of the AST

Translates from IR to Doc representation in one-to-one fashion

No simplifications

No whitespace formatting

Type annotates everything

Methods

dumpy :: a -> Doc ann