Front.Ast

Description

Sslang abstract syntax tree.

Documentation

newtype Program

A complete program: a list of top-level definitions.

Constructors

Program [TopDef]

data TopDef

A top-level definition.

Constructors

TopDef Definition
Bind a (data) value to a variable

TopType TypeDef
Define an algebraic data type

TopCDefs String
Inlined block of C definitions

TopExtern ExternDecl
Declare external symbol for FFI

data ExternDecl

Associate a type with a symbol

Constructors

ExternDecl Identifier Typ

data TypeDef

An algebraic data type definition.

Constructors

TypeDef

Fields

typeName :: Identifier

The name of the type, e.g., Option

typeParams :: [Identifier]

List of type parameters, e.g., a

typeVariants :: [TypeVariant]

List of variants, e.g., Some, None

data TypeVariant

A type variant, i.e., a data constructor.

Constructors

VariantUnnamed Identifier [Typ]

data Definition

A value definition.

Constructors

DefFn Identifier [Pat] TypFn Expr

DefPat Pat Expr

data Pat

A pattern appearing on the LHS of a definition or match arm

Constructors

PatWildcard
Match anything, i.e., _

PatId Identifier
Variable or data constructor, e.g., v or Some

PatLit Literal
Literal match, e.g., 1

PatAs Identifier Pat
Pattern alias, e.g., a @ pat

PatTup [Pat]
Match on a tuple, e.g., (pat, pat)

PatApp [Pat]
Match on multiple patterns, e.g., Some a

PatAnn Typ Pat
Match with type annotation, e.g., pat: Type

data TypFn

Function type annotation

Constructors

TypReturn TypAnn

TypProper TypAnn

TypNone

type TypAnn = Typ

TODO: type classes

data Typ

A type definition

Constructors

TCon Identifier

TApp Typ Typ

TTuple [Typ]

TArrow Typ Typ

data Expr

An expression

Constructors

Id Identifier

Lit Literal

Apply Expr Expr

Lambda [Pat] Expr

OpRegion Expr OpRegion

NoExpr

Let [Definition] Expr

While Expr Expr

Loop Expr

Par [Expr]

IfElse Expr Expr Expr

After Expr Expr Expr

Assign Expr Expr

Constraint Expr TypAnn

Wait [Expr]

Seq Expr Expr

Break

Match Expr [(Pat, Expr)]

CQuote String

CCall Identifier [Expr]

Tuple [Expr]

ListExpr [Expr]

Last Expr

Now

data OpRegion

An operator region: a flat list of alternating expressions and operators that is initially parsed flat but will be restructured into a tree by the operator precedence parser.

Constructors

NextOp Identifier Expr OpRegion

EOR

data Literal

A literal

Constructors

LitInt Integer

LitString String

LitRat Rational

LitChar Char

LitEvent

data Fixity

Fixity declaration for binary operators.

Constructors

Infixl Int Identifier

Infixr Int Identifier

foldApp :: Expr -> [Expr] -> Expr

Apply a function to zero or more arguments.

Suppose we have as source code type Color = RGB Int Int Int ... let x = RGB 203 200 100 and rgb = Id RGB r = Lit (A.LitInt 203) g = Lit (A.LitInt 200) b = Lit (A.LitInt 100), then foldApp rgb [r, g, b] returns Apply (Apply (Apply (Id RGB) (Lit (LitInt 100))) (Lit (LitInt 200))) (Lit (LitInt 203))

collectTApp :: Typ -> (Typ, [Typ])

Collect a type application into the type constructor and its arguments.

collectApp :: Expr -> (Expr, [Expr])

Collect a curried application into the function and its list of arguments.

collectPApp :: Pat -> (Pat, [Pat])

Collect a pattern application into the destructor and arguments.

getTopDataDef :: TopDef -> Maybe Definition

Unwrap a (potential) top-level data definition.

getTopTypeDef :: TopDef -> Maybe TypeDef

Unwrap a (potential) top-level type definition.

getTopCDefs :: TopDef -> Maybe String

Unwrap a (potential) top-level C inline block.

getTopExtern :: TopDef -> Maybe ExternDecl

Unwrap a (potential) top-level external definition.

getTops :: [TopDef] -> ([TypeDef], [String], [ExternDecl], [Definition])

Unzip a list of top-level declarations into their counterparts.