IR.InsertRefCounting

Description

This inserts dup and drop primitives according to a caller dup, callee drop policy. The value returned by a function should be passed back referenced (ownership transfers from the callee back to the caller).

The dup : a -> a primitive behaves like the identity function, evaluating and returning its first argument and increasing the reference count on the result. It is meant to be wrapped around function arguments.

The drop : a -> b -> a primitive evaluates and returns its first argument. It decrements the reference count to its second argument after it has evaluated its first argument. It is meant to be wrapped around function bodies that need to use and then de-reference their arguments.

Thus, something like

add a b = a + b

becomes

add a b =
  drop
    (drop
       ((dup a) + (dup b))
       b)
    a

Arguments a and b to the + primitive are duplicated and the result of + is duplicated internally, so add does not need to duplicate its result. Both arguments a and b are dropped.

Try running sslc --dump-ir-final on an example to see the inserted dup and drop constructs.

Our approach was inspired by Perceus https://www.microsoft.com/en-us/research/publication/perceus-garbage-free-reference-counting-with-reuse/

The external interface

insertRefCounting :: Program Type -> Pass (Program Type)

Insert dup and drop primitives throughout a program

Applies insertTop to the program's definitions

Module internals, not intended for use outside this module

type Fresh = StateT Int Pass

Monad for creating fresh variables: add an Int to the pass

getFresh :: String -> Fresh VarId

Create a fresh variable name with the given suffix

Creates a new variable ID of the form "anon1," "anon2," etc. followed by the supplied suffix

makeDup

Arguments

:: Expr Type
The variable to duplicate and return

-> Expr Type
The dup call

Make a dup primitive that returns the type of its argument

makeDrop

Arguments

:: Expr Type
The expression to evaluate and return

-> Expr Type
The variable to drop afterwards

-> Expr Type
The drop call

Make a drop primitive with unit type

insertTop :: (VarId, Expr Type) -> Fresh (VarId, Expr Type)

Insert referencing counting for top-level expressions

Applies insertExpr to a top-level delcaration

insertExpr :: Expr Type -> Fresh (Expr Type)

Insert reference counting into an expression

This is the main workhorse of this module.

Literals are unchanged, e.g.,

42

remains

42

Data constructors are unchanged, e.g.,

True

remains

True

because they are functions whose results are returned with an existing reference

A variable reference becomes a call to dup because it introduces another reference to the named object, e.g.,

v

becomes

dup v

Application recurses (inserts dups and drops) on both the function being applied and its argument e.g.,

add x y

becomes

(dup add) (dup x) (dup y)

Primitive function application inserts dups and drops on its arguments, e.g.,

(+) x y

becomes

(+) (dup x) (dup y)

Let introduces new names whose values are dropped after the body is evaluated; let _ = are given names so they can be dropped.

let a = Foo 42
    _ = a
17

becomes

let a = Foo 42
drop
  (let anon1_underscore = dup a
   drop
     17
     anon1_underscore)
  a

Nested Lambda expressions are handled by collecting them into a single expression with multiple arguments, adding dups and drops to the body, and adding drops around the body for each argument (which the caller should have duped)

add a b = a + b

desugars to

add = fun a (fun b (a + b))

and becomes

add = fun a (
        fun b (
           drop (
              drop (
                (dup a) + (dup b)
              ) b
           ) a

Matches that operate on a variable are modified by inserting dups and drops into the arms (but not the scrutinee)

match v
  Foo x = x + 1
  Bar = 42

desugars to

match v
  Foo pat_anon0 = let x = pat_anon0
                    x + 1
  Bar = 42

then becomes

match v
  Foo pat_anon0 = drop
                    dup pat_anon0
                    let x = dup pat_anon0
                    drop (
                       dup x + 1
                    ) x
                  ) pat_anon0
  Bar = 42

Matches that scrutinize an expression lift the scrutinee into a let then insert dups and drops on the whole thing

  match add x y
    10 = 5
    _ = 3

becomes

  let anon0_scrutinee = (dup add) (dup x) (dup y)
  drop (
    match anon0_scrutinee
      10 = 5
      _ = 3
  ) anon0_scruitinee

insertAlt :: (Alt, Expr Type) -> Fresh (Alt, Expr Type)

Insert dups and drops into pattern match arms

The body of default and literal patterns is simply recursed upon.

Every named variable in a pattern is duped and dropped, e.g.,

match v
  Foo x = expr

becomes

match v
  Foo _anon1 = drop
                 (let x = dup _anon1
                  expr
               ) x